move tool, and fix most everything
