+++ lib/langchain/agent/chain_of_thought_agent/chain_of_thought_agent.rb
@@ -92,7 +92,7 @@ module Langchain
           question: question,
           tool_names: "[#{tools.join(", ")}]",
           tools: tools.map do |tool|
-            "#{tool}: #{Tool.const_get(Tool::Base::TOOLS[tool]).const_get(:DESCRIPTION)}"
+            "#{tool}: #{Langchain::Tool.const_get(Tool::Base::TOOLS[tool]).const_get(:DESCRIPTION)}"
           end.join("\n")
         )
       end
