+++ spec/langchain/tool/calculator_spec.rb
@@ -2,7 +2,7 @@
 
 require "eqn"
 
-RSpec.describe Tool::Calculator do
+RSpec.describe Langchain::Tool::Calculator do
   describe "#execute" do
     it "calculates the result" do
       expect(subject.execute(input: "2+2")).to eq(4)
@@ -11,7 +11,7 @@ RSpec.describe Tool::Calculator do
     it "calls Serp API when eqn throws an error" do
       allow(Eqn::Calculator).to receive(:calc).and_raise(Eqn::ParseError)
 
-      expect(Tool::SerpApi).to receive(:execute_search).with(input: "2+2").and_return({answer_box: {to: 4}})
+      expect(Langchain::Tool::SerpApi).to receive(:execute_search).with(input: "2+2").and_return({answer_box: {to: 4}})
 
       subject.execute(input: "2+2")
     end
